requirements.txt
Django>=4.2
djangorestframework
django-crispy-forms
django-countries
python-dotenv
reportlab
